+++ docs/cda-logical-model/en/assets/js/direction-management.js
@@ -1,0 +1,6 @@
+if (document.documentElement.dir=='rtl') {
+  var r = document.querySelector(':root');
+
+  r.style.setProperty('--ig-left', 'right');    
+  r.style.setProperty('--ig-right', 'left');
+}