+++ docs/cda-logical-model/en/assets/js/flags.js
@@ -1,0 +1,7 @@
+$('#select_flg').click(function(event){
+  $("#flaglist").show();
+  event.stopPropagation();
+});
+$(document.body).click(function (){
+  $("#flaglist").hide();
+});